Statistics added
Only in minimal-version so the players can see what goea on on their
Server. It shows deaths, block-places, block-destroys, joins; It is
working correctly, but at least a Database with the ciorrect setup is
needed. Almost everything is controlled over the minecraft user. I'm
uploading this, in order to prevent code loose when i'm flashing my
bios.